Extend unittest to include convert_table_to_clip
