--- test/nicos_ess/loki/test_table_to_clipboard_text.py
@@ -1,5 +1,0 @@
-from nicos_ess.utilities.table_utils import convert_table_to_clipboard_text
-
-
-def test_():
-    pass
